requirements.txt
# Framework
Appium-Python-Client
selenium

# allure reporing
allure-behave

